requirements.txt
google-cloud-storage
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
